Expose add_path hook and fix EXPLAIN infinite recursion
The add_path hook is used to only add paths as required by the hints and
prevents many issues that would need to be dealt with otherwise. Using
add_path, the default join search can be re-used, ensuring that
cardinality estimates, parallelization behaviour, etc. all stay the same
compared to a vanilla optimization (for some reason, cardinality
estimates in Postgres do not seem to be associative).

The EXPLAIN bug occurred, when no query plan could be found during
EXPLAINing a hinted query. In this case, the temporary optimizer hooks were
not cleaned up properly, which resulted in infinite recursion.