[HINTING] Fixed being unable to combine mergejoin and materialization
As it turns out, Postgres uses two different techniques to insert
materialization nodes. First, it can explicitly create materialized
versions of the input paths when building a join path. This strategy is
used for nested-loop joins. Second - and this is what broke the merge
join behavior - it might mark the merge join to materialize its input
during the cost estimation. However, PG never actually builds a
materialize path in this case. Therefore, our path acceptance/rejection
logic failed. Instead, the optimizer inserts the materialize *plan*
on-the-fly when transforming the selected access path into a plan.

The solution is somewhat hacky: instead of relying entirely on path
acceptance/rejection, we manipulate the materialization switch directly
for merge joins, depending on the supplied hints.