[HINTING] Fix interaction of parallel hints and join operators
As it turns out, the planner does not simply take input paths for
parallel joins from the partial_pathlist of the inner RelOptInfos.
In reality, this is only done for the *outer* input. The *inner* input
however was selected directly from the pathlist (or at least the
parallel-safe paths).

We did not account for this and rejected all non-placeholder input paths
for the normal pathlist if we had to produce a parallel (sub)plan. This
commit changes this behavior and simply accepts all sequential paths
that satisfy our hints. We will sort out the sequential paths anyway as
soon as we encouter a gather path.