[HINTING] Fix detecting too many paths as placeholders and pruning them
This is a follow up to commit 18d631034e62baafa976b406725ad27ac8ab231d.
Before, we rejected all sequential paths if we had to produce a parallel
plan. The former commit changed this behavior since parallel joins are
actually constructed by combining a partial outer path with a
sequential, parallel-safe inner path.

However, that commit did not account for the fact that the pruning of
placeholders also performs checks for hint compatibility. Specifically,
whenever we accept a new sequential path, we prune all paths that are
compatible with our hints and only act as placeholders to prevent empty
pathlists for the RelOptInfo. But this check still included a
parallelization-specific check resulting in some paths being pruned that
are actually compatible with our hints (since they function as
parallel-safe inner inputs).

The fix for this situation is actually quite straightforward: we do not
check for parallelization unless we encouter a parallel subpath or we
are dealing with the final plan.